% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{plotRaster}
\alias{plotRaster}
\title{plotRaster}
\usage{
plotRaster(
  input,
  assay_name = NULL,
  feature_name = "sum",
  factor_levels = NULL,
  showLegend = TRUE,
  plotTitle = NULL,
  showAxis = FALSE,
  ...
)
}
\arguments{
\item{input}{\code{SpatialExperiment}: Input data represented as a 
\code{SpatialExperiment}. The given \code{SpatialExperiment} is assumed to have 
an \code{assay} slot containing a features-by-observations matrix as \code{dgCmatrix} 
or \code{matrix} and a \code{spatialCoords} slot containing spatial x,y coordinates of 
observations as matrix array. The features-by-observations matrix is assumed to 
have either genes or cell types as features and pixels as observations. Further, 
x,y coordinates are assumed to be stored in column 1 and 2 of \code{spatialCoords}.}

\item{assay_name}{\code{character}: Name of the assay slot of the input that 
you want to visualize. If no argument is given, the first assay of the input 
would be visualized. This argument is useful when you have multiple assays stored 
in the input, and you want to visualize a specific assay. Default is NULL.}

\item{feature_name}{\code{character}: Name of the feature in the input that you 
want to visualize. This argument is useful when you want to specify a feature you 
want to visualize. You can also use "sum" to visualize sum of all feature 
values per observation or "mean" to visualize mean of all feature values per observation. 
Default is "sum".}

\item{factor_levels}{\code{character} or \code{numeric} or \code{factor}: An 
optional vector to convert and plot the input data as \code{factor}. This argument 
is useful if you want to plot categorical/ordinal variables, such as binarized 
occurrence of a specific cell type. \code{factor_levels} is fed into \code{levels} 
argument of the \code{factor} function in base R. Default is NULL.}

\item{showLegend}{\code{logical}: Boolean to show the legend. Default is TRUE.}

\item{plotTitle}{\code{character}: An optional argument to add a title to the 
resulting plot. Default is NULL.}

\item{showAxis}{\code{logical}: Boolean to show axis title, texts, and ticks. 
Default is FALSE.}

\item{...}{Additional parameters to pass to \code{ggplot2::scale_fill_viridis_c} 
if no argument is provided to \code{factor_levels} or \code{ggplot2::scale_fill_viridis_d} 
if a vector is provided to \code{factor_levels}. If you wish to use other color 
maps, we recommend overriding the resulting \code{ggplot} object.}
}
\value{
The output is returned as a \code{ggplot} object, where the input is 
visualized as \code{ggplot2::geom_tile}. Coloring of each tile/pixel represent 
the corresponding values of summarized (sum or mean) or specific feature (e.g. 
rasterized gene expression) per observation (pixel). The x,y coordinates represent 
the pixel centroids.
}
\description{
Function based on \code{ggplot2::geom_tile} to visualize a rasterized spatial omics dataset represented as a \code{SpatialExperiment} object.
}
