% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{rasterizeSparseMatrix}
\alias{rasterizeSparseMatrix}
\title{rasterizeSparseMatrix}
\usage{
rasterizeSparseMatrix(data, pos, resolution = 100, fun = "mean")
}
\arguments{
\item{data}{\code{dgCmatrix}: Feature x observation matrix represented as a 
\code{dgCmatrix} object. Features can be genes or cell types. In the case of 
features being cell types, this matrix is assumed to be a sparse model matrix 
with rows as cell types and coloumns as cell IDs.}

\item{pos}{\code{matrix}: Spatial x,y coordinates of observations stored as a 
matrix array. Further, x,y coordinates are assumed to be stored in column 1 
and 2 of \code{spatialCoords}.}

\item{resolution}{\code{integer}: Resolution or side length of each pixel. 
The unit of this parameter is assumed to be the same as the unit of spatial 
coordinates of the input data.}

\item{fun}{\code{character}: If "mean", pixel value for each pixel 
would be the average of gene expression for all cells within the pixel. If 
"sum", pixel value for each pixel would be the sum of gene expression for all 
cells within the pixel.}
}
\value{
The output is returned as a \code{list} containing rasterized feature 
x observation matrix as \code{dgCmatrix}, spatial x,y coordinates of pixel 
centroids as \code{matrix}, and \code{data.frame} containing cell IDs of cells 
that were aggregated in each pixel.
}
\description{
Function to rasterize a given input sparse matrix based on a given position matrix.

This function assumes that inputs are provided as a \code{dgCmatrix} object 
and \code{matrix} for data and position, respectively.
}
