% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{rasterizeMatrix}
\alias{rasterizeMatrix}
\title{rasterizeMatrix}
\usage{
rasterizeMatrix(
  data,
  pos,
  bbox,
  resolution = 100,
  fun = "mean",
  n_threads = 1,
  BPPARAM = NULL
)
}
\arguments{
\item{data}{\code{dgCmatrix} or \code{matrix}: Feature x observation matrix represented as a 
\code{dgCmatrix} or \code{matrix} object. Features can be genes or cell types. In the case of 
features being cell types, this matrix is assumed to be a sparse model matrix 
with rows as cell types and columns as cell IDs.}

\item{pos}{\code{matrix}: Spatial x,y coordinates of observations stored as a 
matrix array. Further, x,y coordinates are assumed to be stored in column 1 
and 2 of \code{spatialCoords}.}

\item{bbox}{\code{bbox}: Bounding box for rasterization defined by a numeric 
vector of length four, with xmin, ymin, xmax and ymax values.}

\item{resolution}{\code{integer} or \code{double}: Resolution or side length of each pixel. 
The unit of this parameter is assumed to be the same as the unit of spatial 
coordinates of the input data.}

\item{fun}{\code{character}: If "mean", pixel value for each pixel 
would be the average of gene expression for all cells within the pixel. If 
"sum", pixel value for each pixel would be the sum of gene expression for all 
cells within the pixel.}

\item{n_threads}{\code{integer}: Number of threads for parallelization. Default = 1. 
Inputting this argument when the \code{BPPARAM} argument is missing would set parallel 
exeuction back-end to be \code{BiocParallel::MulticoreParam(workers = n_threads)}. 
We recommend setting this argument to be the number of cores available 
(\code{parallel::detectCores(logical = FALSE)}). If \code{BPPARAM} argument is 
not missing, the \code{BPPARAM} argument would override \code{n_threads} argument.}

\item{BPPARAM}{\code{BiocParallelParam}: Optional additional argument for parallelization. 
This argument is provided for advanced users of \code{BiocParallel} for further 
flexibility for setting up parallel-execution back-end. Default is NULL. If 
provided, this is assumed to be an instance of \code{BiocParallelParam}.}
}
\value{
The output is returned as a \code{list} containing rasterized feature 
x observation matrix as \code{dgCmatrix} if data was given as \code{dgCmatrix} 
and as \code{matrix} if data was given as \code{matrix}, spatial x,y coordinates of pixel 
centroids as \code{matrix}, and \code{data.frame} containing cell IDs of cells 
that were aggregated in each pixel.
}
\description{
Function to rasterize a given input matrix (both dense or sparse) based on a given position matrix.

This function assumes that inputs are provided as a \code{dgCmatrix} 
or \code{matrix} for data and \code{matrix} for position.
}
